+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add `OptionalCodec`
+- Add `DataType::Optional`
+- Implement `Element` for `Option<T>` where `T: Element`
+- Implement `ElementOwned` for `Option<T>` where `T: ElementOwned`
+- Implement `ElementFixedLength` for `Option<T>` where `T: ElementFixedLength`
+
 ## [0.22.5] - 2025-10-26
 
 ### Added